Make the grid endless, optimize life
Previously the grid was given a fixed size.
Now it scales and moves itself to always show all living cells.

Previously all cells were stored in a vector of vectors
and the simulation checked all of them.
Now living cells are stored in a hash map, and all other cells are dead.